foundry_agents/requirements.txt
azure-ai-projects
python-dotenv